refactor: usePermit2->usePermit2Allowance
